requirements-bot.txt
# Зависимости Telegram-бота (aiogram + HTTP-клиент).
aiogram>=3.4,<4.0
aiohttp>=3.9,<4.0
requests>=2.31.0
redis>=5.0.0
psycopg2-binary>=2.9,<3.0
beautifulsoup4>=4.12
# Зависимости Telegram-бота.
